`usage()`: moved from `utility.c` to `ts-warp.c`

+++ logfile.h
@@ -24,10 +24,9 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE. */
 
 /* ------------------------------------------------------------------------------------------------------------------ */
 #include <stdio.h>
-#include <stdint.h>
 
 /* -- Global variables ---------------------------------------------------------------------------------------------- */
-extern uint8_t loglevel;
+extern int loglevel;
 extern FILE *lfile;
 
 /* -- Log verbosity levels ------------------------------------------------------------------------------------------ */
